created separate wifi class, but lcd screen WifiStrength bricks lcd screen
