+++ src/station-lcd-interface/index.js
@@ -20,7 +20,7 @@ import { HostnameTask } from "./tasks/hostname-task.js"
 import { InternetTask } from "./tasks/internet-task.js"
 import { QaqcRequest } from './tasks/qaqc-task.js'
 import { BashUpdateTask } from './tasks/bash-update.js'
-import { WifiStrength } from './tasks/wifi-strength.js'
+// import { WifiStrength } from './tasks/wifi-strength.js'
 import ButtonMap from './button-map.js'
 
 // Require Statements
@@ -43,7 +43,7 @@ const host = 'http://localhost:3000'
 */
 
 let items = new MenuItem("main", null, [
-  new MenuItem("Wifi Strength", new WifiStrength(), []),
+  // new MenuItem("Wifi Strength", new WifiStrength(), []),
 
   new MenuItem("File Transfer", null, [
     new MenuItem("Mount Usb", new MountUsbTask(host), []),
